refactor: フォルダ構成の整理とVueファイルの移動。
