+++ src/layouts/default.vue
@@ -141,7 +141,7 @@ import StrageDatasource from '@/infrastructure/strage/StrageDatasource'
 import ClientDownloadRepository from '@/domain/client/ClientDownloadRepository'
 import ClientDownloadTransfar from '@/infrastructure/client/ClientDownloadTransfar'
 import ProductSelectorDialog from '@/components/ProductSelectorDialog.vue'
-import ProductImportDialog from  '~/components/ProductImportDialog.vue'
+import ProductImportDialog from  '@/components/product/import/ProductImportDialog.vue'
 import LocalStrageInitializeDialog from '@/components/LocalStrageInitializeDialog.vue'
 import LocalStrageDestroyDialog from '@/components/LocalStrageDestroyDialog.vue'
 import ProductImportService from '~/application/service/product/import/ProductImportService'
